[refactor] Extracted `HasTable` and `HasForm` from the Houston::Extensions::View (2m)
